requirements.txt
Appium-Python-Client>=0.28
pytest
pytest-html